requirements.txt
ttkbootstrap>=1.10
Pillow>=10.0
PyMuPDF>=1.24
ezdxf>=0.19
matplotlib>=3.7
